CKS-275 Refactor data layer
